refactor: API Response 객체에서 API Error Response를 분리
