+++ src/main/java/es/princip/getp/infra/dto/response/ApiResponse.java
@@ -2,8 +2,6 @@ package es.princip.getp.infra.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import es.princip.getp.infra.exception.ErrorCode;
-import es.princip.getp.infra.exception.ErrorDescription;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,19 +23,6 @@ public class ApiResponse {
     public static ResponseEntity<ApiSuccessResult<?>> success(final HttpStatus status) {
         return ResponseEntity.status(status).body(ApiResponse.body(status));
     }
-
-    public static ResponseEntity<ApiErrorResult> error(final ErrorCode errorCode) {
-        return ResponseEntity.status(errorCode.status())
-            .body(new ApiErrorResult(errorCode));
-    }
-
-    public static ResponseEntity<ApiErrorResult> error(
-        final HttpStatus status,
-        final ErrorDescription... descriptions
-    ) {
-        return ResponseEntity.status(status)
-            .body(new ApiErrorResult(status.value(), descriptions));
-    }
     
     @JsonInclude(Include.NON_NULL)
     @Getter
@@ -53,23 +38,6 @@ public class ApiResponse {
         private ApiSuccessResult(final int status) {
             this.status = status;
             this.data = null;
-        }
-    }
-
-    @JsonInclude(Include.NON_NULL)
-    @Getter
-    public static class ApiErrorResult {
-        private final int status;
-        private final ErrorDescription[] errors;
-
-        private ApiErrorResult(final ErrorCode errorCode) {
-            this.status = errorCode.status().value();
-            this.errors = new ErrorDescription[] { errorCode.description() };
-        }
-
-        private ApiErrorResult(final int status, final ErrorDescription... descriptions) {
-            this.status = status;
-            this.errors = descriptions;
         }
     }
 }
